// web/src/api/generated/model/eventIndex200Item.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Les Canetons API
 * OpenAPI spec version: 0.0.1
 */

export type EventIndex200Item = {
  id: number;
  date: string;
  title: string;
  startTime: string;
  endTime: string;
  location: string;
  /** @nullable */
  attire: string | null;
  weekend: number;
  /** @nullable */
  response: string | null;
};